TPASPKT-1530 Migrate path handling to `importlib`
`pkg_resources` module installed via setuptools is deprecated and might
be removed as early as this year. `importlib.resources` is a standard
library module available starting with Python 3.7 that does what we
need.

The minimum Python version requirement is also bumped from 3.6 to 3.9.
Technically we could also support 3.7, but 3.9 is used for consistency
with other DPRES projects.