TPASPKT-1510 Allow external file format data
Allow the user to load external file format JSON data and provide it to
`file_formats` function.

The main reason is to allow library users to download JSON data from the
internet and provide it to the library. This is more suitable for
automated environments where we want the file format data to remain
up-to-date without requiring Python package updates, which require
manual effort and oversight.