TPASPKT-1608 Replace version with setuptools-scm
Replace hardcoded version number with setuptools-scm, same as what
several other of our Python packages already use.

A common problem with the hardcoded version number is that we forget to
update it in addition to the changelog and git tag, meaning any Python
tools like `pip` will not detect the new version correctly.
setuptools-scm also applies dirty tags to the version number, which can
improve metadata provenance in some situations (eg. a customer uses a patched
version of dpres-file-formats in their environment).